Adopt existing channel on create (#113)
* Adopt existing channel on create
If a new 'adopt_existing_channel' attribute is set 'true' (default 'false'
for backwards compatibility), an existing channel with the same name
and type will be 'adopted' by the terraform provider, and unarchived
if necessary, to serve as the newly 'created' channel.
This brings parity with the existing provider functionality of archiving
or abandoning channels on terraform resource destroy. Now you can do
a 'terraform apply', 'terraform destroy', and 'terraform apply' again
without errors.

* Make sure api user is in the channel before doing channel
membership operations (needed when adopting existing unarchived
channels)

* Update to terraform-plugin-log v0.4.0
The upstream provider updated the sdk version, which pulled in this
new version. The new version has breaking API changes to logging to
make it more 'structured'.

* Update documentation on conversation api.

* Update go.mod and go.sum for new deps

* Update go.mod to version 0.7.0 of terraform-plugin-log

* added newline

* Made changes to satisfy pedantic linter.

* Address pedantic docs linter whitespace issues.

* Fix test failures
- some channel types (notably private ones) can't be joined.
- ephemeral attributes shouldn't be checked